requirements.txt
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
mccabe==0.5.0
Pillow==3.2.0
pycodestyle==2.0.0
python-dateutil==2.5.3
six==1.10.0
voluptuous==0.8.11
Werkzeug==0.11.9
gunicorn==19.6.0
structlog==16.1.0
